% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resume.R
\name{resume}
\alias{resume}
\title{Launch Resume in the default browser}
\usage{
resume(app = "")
}
\arguments{
\item{app}{Choose one}
}
\description{
Launch Resume in the default browser
}
\details{
...
}

